+++ test/services/stake_and_mint/ProgressService/update.test.ts
@@ -3,14 +3,14 @@ import Web3 from 'web3';
 
 import { interacts } from '@openst/mosaic-contracts';
 
-import { AUXILIARY_GAS_PRICE, ORIGIN_GAS_PRICE } from '../../../src/Constants';
-import Message from '../../../src/models/Message';
-import GatewayRepository from '../../../src/repositories/GatewayRepository';
-import { MessageStatus } from '../../../src/repositories/MessageRepository';
-import ProgressService from '../../../src/services/stake_and_mint/ProgressService';
-import Utils from '../../../src/Utils';
-import SpyAssert from '../../test_utils/SpyAssert';
-import StubData from '../../test_utils/StubData';
+import { AUXILIARY_GAS_PRICE, ORIGIN_GAS_PRICE } from '../../../../src/Constants';
+import Message from '../../../../src/models/Message';
+import GatewayRepository from '../../../../src/repositories/GatewayRepository';
+import { MessageStatus } from '../../../../src/repositories/MessageRepository';
+import ProgressService from '../../../../src/services/stake_and_mint/ProgressService';
+import Utils from '../../../../src/Utils';
+import SpyAssert from '../../../test_utils/SpyAssert';
+import StubData from '../../../test_utils/StubData';
 
 describe('ProgressService.update()', () => {
   let coGatewaySpy: any;
