+++ test/services/stake_and_mint/ProveGateway/proveGateway.test.ts
@@ -5,15 +5,15 @@ import Web3 from 'web3';
 import { interacts } from '@openst/mosaic-contracts';
 import { ProofGenerator } from '@openst/mosaic-proof';
 
-import { AUXILIARY_GAS_PRICE } from '../../../src/Constants';
-import Gateway from '../../../src/models/Gateway';
-import GatewayRepository from '../../../src/repositories/GatewayRepository';
-import {MessageDirection, MessageRepository} from '../../../src/repositories/MessageRepository';
-import ProveGatewayService from '../../../src/services/stake_and_mint/ProveGatewayService';
-import Utils from '../../../src/Utils';
-import assert from '../../test_utils/assert';
-import SpyAssert from '../../test_utils/SpyAssert';
-import StubData from '../../test_utils/StubData';
+import { AUXILIARY_GAS_PRICE } from '../../../../src/Constants';
+import Gateway from '../../../../src/models/Gateway';
+import GatewayRepository from '../../../../src/repositories/GatewayRepository';
+import {MessageDirection, MessageRepository} from '../../../../src/repositories/MessageRepository';
+import ProveGatewayService from '../../../../src/services/stake_and_mint/ProveGatewayService';
+import Utils from '../../../../src/Utils';
+import assert from '../../../test_utils/assert';
+import SpyAssert from '../../../test_utils/SpyAssert';
+import StubData from '../../../test_utils/StubData';
 
 describe('ProveGatewayService.proveGateway()', (): void => {
   const originWeb3 = new Web3(null);
